% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{add_ranks}
\alias{add_ranks}
\title{Agglomerate and merge a phyloseq object}
\usage{
add_ranks(ps = ps, ranks = c(rank_names(ps)),
  replace_ASV_names = TRUE, verbose = TRUE)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains a \code{\link[phyloseq]{tax_table}}.}

\item{ranks}{A vector of taxonomic rank names found in \code{\link[phyloseq]{rank_names}}. Defaults to all available ranks (\code{Phylum} through \code{Genus}).}

\item{replace_ASV_names}{Logical. If TRUE, ASVs will be named with a unique and the lowest taxonomic rank available.}

\item{verbose}{Logical. If FALSE, messages will be supressed.}
}
\value{
This function returns the input \code{phyloseq} object with additional taxa that are the
agglomerated taxa at the taxonomic levels provided.
}
\description{
Agglomerates by \code{rank_names()} and then merges them into a phyloseq object.
}
\examples{
ps <- add_ranks(ps, ranks = c("Genus", "Family"))
}
