% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{calculate_ACE}
\alias{calculate_ACE}
\title{Estimate the A.C.E. model from a phyloseq object}
\usage{
calculate_ACE(ps, zygosity, dz, family_id, verbose = TRUE)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains \code{\link[phyloseq]{sample_data}}.}

\item{zygosity}{The name of the column in sample_data() that contains the \code{zygosity} variable (either MZ or DZ).}

\item{dz}{The level in the \code{zygosity} variable corresponding to dyzogitic twins.}

\item{family_id}{The name of the column in sample_data() that contains the \code{family_id} variable. In other words, both members of a twin pair should have the same value in this column.}

\item{verbose}{Logical. If FALSE, messages will be supressed.}
}
\value{
This function returns a \code{data.frame} with A.C.E. values (and confidence intervals) for each ASV.
}
\description{
Estimates the A.C.E. model of heritabily (with confidence intervals) from a phyloseq object using the \code{\link[mets]{twinlm}} function.
}
\examples{
df <- calculate_ACE(ps, zygosity = "Zygosity", dz = "DZ", family_id = "Family_ID", verbose = TRUE)
}
