% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{normalize_for_heritability}
\alias{normalize_for_heritability}
\title{Normalize count data in a phyloseq object}
\usage{
normalize_for_heritability(ps, covariates = c(),
  normalization_plots = FALSE, verbose = TRUE)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains \code{\link[phyloseq]{sample_data()}}.}

\item{covariates}{A vector of covariates found in \code{\link[phyloseq]{sample_data}}. Defaults to `x ~ 1`.}

\item{normalization_plots}{Logical. If TRUE, creates a directory and saves plots of ASV distributions pre- and post-normalization. Defaults to FALSE.}

\item{verbose}{Logical. If FALSE, messages will be supressed.}
}
\value{
This function returns the input \code{phyloseq} object with the \code{otu_table()} counts normalized.
}
\description{
Normalizes count data in a phyloseq object, first by adding a pseudocount (+1), then transforming to relative abundances, then by normalizes via the Box-Cox transformation in the \code{car} package.
}
\examples{
ps <- normalize_for_heritability(ps, covariates = c("Plate", "Age", "Sex"), normalization_plots = TRUE)
}
