% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heritability}
\alias{plot_heritability}
\title{Plots the output from \code{calculate_ACE()}}
\usage{
plot_heritability(df, cutoff, color, facet = TRUE, verbose = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} that is the output from \code{\link[mherit]{calculate_ACE}}.}

\item{cutoff}{The minimum value of \code{A} to display. Defaults to the cutoff at which the top 25\% of taxonomic features are displayed.}

\item{color}{The taxonomic level to color the plot by. Defaults to the first \code{\link[phyloseq]{rank_names}}.}

\item{facet}{Logical. If TRUE, the output plot will be faceted by taxonomic rank.}

\item{verbose}{Logical. If FALSE, messages will be supressed.}
}
\value{
This function returns a \code{ggplot} plot of the taxnomic features sorted by their heritability and optionally faceted.
}
\description{
Plots the output from \code{calculate_ACE()} which contains the A.C.E. model of heritabily (with confidence intervals).
}
\examples{
p <- plot_heritability(df, cutoff = 0.3, color = "Class", facet = TRUE, verbose = TRUE)
p

}
